% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Algorithm.SDM.R
\docType{class}
\name{Algorithm.SDM-class}
\alias{Algorithm.SDM-class}
\title{An S4 class to represent an SDM based on a single algorithm}
\description{
This is an S4 class to represent an SDM based on a single algorithm (including
generalized linear model, general additive model, multivariate adpatative
splines, generalized boosted regression model, classification tree analysis,
random forest, maximum entropy, artificial neural network, and support vector
machines). It can be obtain with \code{\link{modelling}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the SDM (by default Species.SDM).}

\item{\code{projection}}{raster. Habitat suitability map of the SDM.}

\item{\code{evaluation}}{data frame. Evaluation of the SDM (threshold, AUC, omission
rate, sensitivity, specificity, correct proportion and Kappa).}

\item{\code{variable.importance}}{data frame. Relative percentage of importance for each variable used in the SDM.}

\item{\code{data}}{data frame. Data used to realized the SDM}

\item{\code{parameters}}{data frame. Parameters used to realized the SDM}
}}
\seealso{
\linkS4class{Ensemble.SDM} an S4 class for ensemble SDMs,
 and \linkS4class{Stacked.SDM} an S4 class for SSDMs.
}

