% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ensemble.SDM.R
\docType{class}
\name{Ensemble.SDM-class}
\alias{Ensemble.SDM-class}
\title{An S4 class to represent an ensemble SDM}
\description{
This is an S4 class to represent an ensemble SDM from multiple algorithms
(including generalized linear model, general additive model, multivariate
adpatative splines, generalized boosted regression models, classification tree
analysis, random forest, maximum entropy, artificial neural network, and
support vector machines). It can be obtain with
\code{\link{ensemble_modelling}} or \code{\link{ensemble}}.
}
\section{Slots}{

\describe{
\item{\code{uncertainty}}{raster. Between-algorithm variance map.}

\item{\code{algorithm.correlation}}{data frame. Between-algorithm correlation matrix.}

\item{\code{algorithm.evaluation}}{data frame. Calculated with Pearson's coefficient
 from habitat suitability map.}
}}
\seealso{
\linkS4class{Algorithm.SDM} an S4 class to represent an SDM based on
 a single algorithm, and \linkS4class{Stacked.SDM} an S4 class for SSDMs.
}

