% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Env.R
\docType{data}
\name{Env}
\alias{Env}
\title{A stack of three environmental variables}
\format{
A stack of three rasters: \describe{ \item{RAINFALL}{Annual mean
  rainfall (mm)} \item{TEMPERATURE}{Annual mean temperature (x10 degree
  Celsius)} \item{SUBSTRATE}{Substrate type (categorical variable)} }
}
\usage{
Env
}
\description{
A stack of three 30 arcsec-resolution rasters covering the north part of the
main island of New Caledonia 'Grande Terre'. CRAINFALL and
TEMPERATURE rasters are climatic variables from the WorldClim database, and SUBSTRATE raster is from
the IRD Atlas of New Caledonia (2012) (see reference below).
}
\references{
R.J. Hijmans, C.H. &  Graham (2006) "The  ability  of climate
  envelope  models  to  predict  the  effect  of  climate  change  on species
  distributions." \emph{Global Change Biology} 12:2272-2281
  \url{https://onlinelibrary.wiley.com/doi/full/10.1111/j.1365-2486.2006.01256.x}

  E. Fritsch (2012) "Les sols. Atlas de la Nouvelle-Caledonie (ed. by J.
  Bonvallot, J.-C. Gay and E. Habert)" \emph{IRD-Congres de la
  Nouvelle-Caledonie, Marseille.} 73-76
}
\keyword{datasets}
