% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.Stacked.SDM.R
\docType{methods}
\name{evaluate.Stacked.SDM}
\alias{evaluate.Stacked.SDM}
\title{Evaluate}
\usage{
\S4method{evaluate}{Stacked.SDM}(obj, ...)
}
\arguments{
\item{obj}{Stacked.SDM. SSDM to evaluate}
}
\value{
SSDM evaluation in a data.frame
}
\description{
Evalaution of SSDM floristic composition with Pottier et al, 2013 method (see reference below)
}
\examples{

## Not run:
# Loading data
data(Env)
data(Occurrences)
# SSDM building
SSDM <- stack_modelling(c('CTA', 'SVM'), Occurrences, Env, rep = 1,
                       Xcol = 'LONGITUDE', Ycol = 'LATITUDE',
                       Spcol = 'SPECIES')

# Evaluation
evaluate(SSDM)

## End(Not run)

}
\references{
Pottier, J., Dubuis, A., Pellissier, L., Maiorano, L., Rossier, L., Randin, C.
F., … Guisan, A. (2013). The accuracy of plant assemblage prediction from
species distribution models varies along environmental gradients. Global
Ecology and Biogeography, 22(1), 52–63.
https://doi.org/10.1111/j.1466-8238.2012.00790.x
}
