% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_model.R
\name{load.model}
\alias{load.model}
\alias{load_enm}
\alias{load_stack}
\title{Function to load ensemble SDMs and SSDMs}
\usage{
load_enm(name, directory = getwd())

load_stack(name = "Stack", directory = getwd(), GUI = F)
}
\arguments{
\item{name}{character. Name of the folder that contains the model to be loaded.}

\item{directory}{character. Path to the directory containing the saved model
folder, by default the current directory.}

\item{GUI}{logical. Don't take that argument into account (parameter for the
  user interface).}
}
\value{
The corresponding SDM object.
}
\description{
Allow to load S4 \linkS4class{Ensemble.SDM} and \linkS4class{Stacked.SDM}
objects saved with their respective save function.
}
\seealso{
\code{\link{save.model}}
}

