% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/load_var.R
\name{load_var}
\alias{load_var}
\title{Load environmental data}
\usage{
load_var(directory = getwd(), files = NULL, format = c(".grd", ".tif",
  ".asc", ".sdat", ".rst", ".nc", ".tif", ".envi", ".bil", ".img"),
  factors = NULL, Norm = T, tmp = T, verbose = T, GUI = F)
}
\arguments{
\item{directory}{character. Directory that contains the environmental data
files.}

\item{files}{character. Files containing the environmental data. If NULL
(default) all files with the precised format present in the directory will
be loaded.}

\item{format}{character. Format used to load environmental data variable files
(including .grd, .tif, .asc, .sdat, .rst, .nc, .tif, .envi, .bil, .img)}

\item{factors}{character. Environmental data which should be considered as
factor variable.}

\item{Norm}{logical. Normalize environmental data.}

\item{tmp}{logical. If true loaded environmental data variable rasters are
read in temporary file avoiding to overload the random access memory. But
beware: if you close R, temporary files will be destructed and you'll need to
reload your environmental datas directly from their original files.}

\item{verbose}{logical. If true allow the function to print text in the
console}

\item{GUI}{logical. Don't take that argument into account (parameter for the
 user interface) !}
}
\value{
A stack containing the environmental rasters (normalized or
 not)
}
\description{
Function to load environmental data from rasters to perform
\code{\link{modelling}}, \code{\link{ensemble_modelling}} or
\code{\link{stack_modelling}}.
}
\examples{
\dontrun{
load.var(directory)
}
}
\seealso{
\code{\link{load_occ}} to load occurences
}

