% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Occurrences.R
\docType{data}
\name{Occurrences}
\alias{Occurrences}
\title{Plant occurrence data frame}
\format{A data frame with 1343 rows and 3 variables: \describe{
  \item{Genus}{Genus of the species}
  \item{Longitude}{Longitude of the observation}
  \item{Latitude}{Latitude of the observation} }}
\usage{
Occurrences
}
\description{
A dataset containing plant occurrences from 7 species of New Caledonia at a 5
km-resolution. Occurrence data come from Noumea Herbarium (NOU) and NC-PIPN
network (see Ibanez et al in references below).
}
\references{
T. Ibanez, J. Munzinger, G. Dagostini, V. Hequet, F. Rigault, T.
  Jaffre, & P. Birnbaum (2014) "Structural and floristic characteristics of
  mixed rainforest in New Caledonia: new data from the New Caledonian Plant
  Inventory and Permanent Plot Network (NC-PIPPN)." \emph{Applied Vegetation
  Science} 17:386-397
  \url{http://www.researchgate.net/profile/Jerome_Munzinger/publication/
  258499017_Structural_and_floristic_diversity_of_mixed_tropical_rain_forest_in_New_Caledonia_new_data_
  from_the_New_Caledonian_Plant_Inventory_and_Permanent_Plot_Network_%28NC-PIPPN%29/links/
  0deec52b8b1996488e000000.pdf}
}
\keyword{datasets}

