% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.model.R
\docType{methods}
\name{plot.model}
\alias{plot,SDM,ANY-method}
\alias{plot,Stacked.SDM,ANY-method}
\alias{plot.model}
\title{Plot all SDMs, ensemble SDMs, and SSDMs}
\usage{
\S4method{plot}{Stacked.SDM,ANY}(x, y, ...)

\S4method{plot}{SDM,ANY}(x, y, ...)
}
\arguments{
\item{x}{Object to pe plot (S4 Algorithm.SDM, Ensemble.SDM or
Stacked.SDM object).}

\item{y,...}{Don't used plot base parameters.}
}
\value{
Open a window with a shiny app rendering all the results in an easy
  way to see all results in one glance.
}
\description{
Allow to plot S4 \linkS4class{Algorithm.SDM},
\linkS4class{Ensemble.SDM} and
\linkS4class{Stacked.SDM} classes objects.
}

