% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{project}
\alias{project}
\alias{project,Algorithm.SDM-method}
\alias{project,MAXENT.SDM-method}
\alias{project,Ensemble.SDM-method}
\title{Project model into a new environment}
\usage{
project(obj, Env, ...)

\S4method{project}{Algorithm.SDM}(obj, Env, ...)

\S4method{project}{MAXENT.SDM}(obj, Env, ...)

\S4method{project}{Ensemble.SDM}(obj, Env, ...)
}
\arguments{
\item{obj}{Object of class Algorithm.SDM, Ensemble.SDM or Stacked.SDM. Model(s) to be projected.}

\item{Env}{Raster stack. Updated environmental rasters to be used for projection.}

\item{...}{}
}
\value{
A raster (Algorithm.SDM), raster stack (Ensemble.SDM), biodiversity map/mean raster (Stacked.SDM)
}
\description{
This is a method to project existing SDMs, ESDMs or SSDMs to a new environment. The function uses any S4 .SDM class object and either returns the object with updated projection slots or only returns the projection as rasters (if minimal.outputs = TRUE)
}
