% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Stacked.SDM.R
\docType{class}
\name{Stacked.SDM-class}
\alias{Stacked.SDM-class}
\title{An S4 class to represent SSDMs}
\description{
This is an S4 class to represent SSDMs of multiple algorithms
(including generalized linear model, general additive model, multivariate
adpatative splines, generalized boosted regression models, classification tree
analysis, random forest, maximum entropy, artificial neural network, and
support vector machines). It can be obtain with \code{\link{stack_modelling}}
or \code{\link{stacking}}.
}
\section{Slots}{

\describe{
\item{\code{name}}{character. Name of the SDM (by default Species.SSDMs).}

\item{\code{diversity.map}}{raster. Local species richness map of the Stacked SDMs}

\item{\code{uncertainty}}{raster. Between-algorithm variance map.}

\item{\code{evaluation}}{data frame. Evaluation of the Stacked SDMs (threshold, AUC, omission
rate, sensitivity, specificity, correct proportion and Kappa).}

\item{\code{variable.importance}}{data frame. Relative percentage of importance for
each variable used in the model.}

\item{\code{algorithm.correlation}}{data frame. Between-algorithm correlation matrix.}

\item{\code{enms}}{list. List with each ensemble SDM used for the stacked SDMs.}

\item{\code{parameters}}{data frame. Parameters used to realized the stacked SDMs.}

\item{\code{algorithm.evaluation}}{data frame. Calculated with Pearson's coefficient
 from habitat suitability map.}
}}
\seealso{
\linkS4class{Ensemble.SDM} an S4 class for ensemble SDMs, and
 \linkS4class{Algorithm.SDM} an S4 class for SDMs.
}

