% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/update.stack.R
\docType{methods}
\name{update,Stacked.SDM-method}
\alias{update,Stacked.SDM-method}
\title{Update a previous SSDM}
\usage{
\S4method{update}{Stacked.SDM}(object, Occurrences, Env, Xcol = "Longitude",
  Ycol = "Latitude", Pcol = NULL, Spname = NULL, name = stack@name,
  save = F, directory = getwd(), thresh = 1001, tmp = F, verbose = T,
  GUI = F, ...)
}
\arguments{
\item{object}{Stacked.SDM. The previously produced SSDM.}

\item{Occurrences}{data frame. New or updated occurrence table (can be
processed first by \code{\link{load_occ}}).}

\item{Env}{raster object. Environnment raster object (can be processed first by
\code{\link{load_var}}).}

\item{Xcol}{character. Name of the column  in the occurence table  containing
Latitude or X coordinates.}

\item{Ycol}{character. Name of the column in the occurence table  containing
Longitude or Y coordinates.}

\item{Pcol}{character. Name of the column in the occurence table specifying
wether a line is a presence or an absence. If NULL presence-only data set is
assumed.}

\item{Spname}{character. Name of the new or updated specie.}

\item{name}{character. Optional name given to the final SSDM produced, by
default it's the name of the previous SSDM.}

\item{save}{logical. If true the model is automatically saved.}

\item{directory}{character. Name of the directory to contain the saved SSDM.}

\item{thresh}{numeric. A single integer value representing the number of equal
interval threshold values between 0 & 1. The higher it is the more accurate
is the threshold but the longer is the modelling evaluation step (see
\code{\link[SDMTools]{optim.thresh}}).}

\item{tmp}{logical. If true the habitat suitability map of each algorithm is
saved in a temporary file to release memory. But beware: if you close R,
temporary files will be destroyed. To avoid any loss you can save your model
with \code{\link{save.model}}.}

\item{verbose}{logical. If true allow the function to print text in the
console.}

\item{GUI}{logical. Don't take that argument into account (parameter for the
user interface).}

\item{...}{additionnal parameters for the algorithm modelling function (see
 details below).}
}
\value{
an S4 \linkS4class{Stacked.SDM} class object viewable with
 \code{\link{plot.model}} function.
}
\description{
Update a previous SSDM with new occurrence data. The function takes in inputs updated
or new occurrence data from one specie, previous envrionmental data,
and an S4 \linkS4class{Stacked.SDM} class object containing a previously
produced SSDM.
}
\examples{
\dontrun{
update(stack, Occurrences, Env, Spname = 'NewSpecie')
}
}
\seealso{
\code{\link{stack_modelling}} to build SSDMs.
}

