% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ensemble.R
\docType{methods}
\name{ensemble}
\alias{ensemble}
\alias{ensemble,Algorithm.SDM-method}
\alias{sum,Algorithm.SDM-method}
\title{Methods to assemble different algorithms in an ensemble model}
\usage{
ensemble(x, ..., name = NULL, ensemble.metric = c("AUC"),
  ensemble.thresh = c(0.75), weight = T, thresh = 1001, uncertainty = T,
  verbose = T, GUI = F)

\S4method{ensemble}{Algorithm.SDM}(x, ..., name = NULL,
  ensemble.metric = c("AUC"), ensemble.thresh = c(0.75), weight = T,
  thresh = 1001, uncertainty = T, verbose = T, GUI = F)

\S4method{sum}{Algorithm.SDM}(x, ..., name = NULL,
  ensemble.metric = c("AUC"), ensemble.thresh = c(0.75), weight = T,
  thresh = 1001, format = T, verbose = T, na.rm = F)
}
\arguments{
\item{x,...}{SDMs. SDMs to be assembled.}

\item{name}{character. Optionnal name given to the final Ensemble.SDM
producted.}

\item{ensemble.metric}{character. Metric used to select SDMs that will be
included in the ensemble model (see details below).}

\item{ensemble.thresh}{numeric. Threshold associated with the metric used to
compute the selection.}

\item{weight}{logical. Choose wether or not you want the SDMs to be weighted
using the mean of the selection metrics.}

\item{thresh}{numeric. Threshold precision parmeter used to compute binary
map, the higher it is the more accurate is the threshold but the longer is
the modelling evaluation step (see \code{\link[SDMTools]{optim.thresh}}).}

\item{uncertainty}{logical. If true uncertainty mapping and algorithms
correlation matrix are computed.}

\item{verbose}{logical. If true allows the function to print text in the
console.}

\item{GUI,format,na.rm}{logical. Don't take those arguments into account (parameters for the
 user interface and sum function) !}
}
\value{
an S4 \linkS4class{Ensemble.SDM} Class object viewable with
 \code{\link{plot.model}} method
}
\description{
This is a method to assemble several algorithms in an ensemble model. It takes
as inputs several S4 \linkS4class{Algorithm.SDM} class objects obtained with
the \code{\link{modelling}} function. It returns an S4
\linkS4class{Ensemble.SDM} class object containing the habitat suitability
map, the binary map, and the uncertainty map (based on the between-algorithm
variance of habitat suitability maps) and all evaluation tables coming with
(model evaluation, algorithms evaluation, algorithms correlation matrix and
variable importance).
}
\details{
ensemble.metric (metric used to select SDMs that will be included in
 the ensemble SDM) can be chosen among : \describe{ \item{AUC}{Area under
 the receiving operative characteristic (ROC) curve} \item{Kappa}{Kappa from
 the confusion matrix} \item{sensitivity}{Sensitivity from the confusion
 matrix} \item{specificity}{Specificity from the confusion matrix}
 \item{prop.correct}{Proportion of correctly predicted occurrences from the
 confusion matrix} }
}
\examples{
\dontrun{
ensemble(GLM1, GLM2, GAM1, GAM2)
}
}
\seealso{
\code{\link{ensemble_modelling}} to build ensemble SDM with multiple
 algorithms.
}

