% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{SSDM package Global User Interface}
\usage{
gui(port = getOption("shiny.port"), host = getOption("shiny.host",
  "127.0.0.1"))
}
\arguments{
\item{port}{char. The TCP port that the application should listen on (see
\code{\link[shiny]{runApp}} for more details).}

\item{host}{char. The IPv4 address that the application should listen on (see
\code{\link[shiny]{runApp}} for more details).}
}
\value{
Open a window with a shiny app to use the SSDM package with an
  user-friendly interface.
}
\description{
User interface of the SSDM package.
}
\details{
If your environmental variables have an important size, you should
  gave enough memory to the interface with the (\code{maxmem} parameter).
}
\examples{
\dontrun{
gui()
}

}
