% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/gui.R
\name{gui}
\alias{gui}
\title{SSDM package Global User Interface}
\usage{
gui()
}
\value{
Open a window with a shiny app to use the SSDM package with an
  user-friendly interface.
}
\description{
User interface of the SSDM package.
}
\details{
If your environmental variables have an important size, you should
  gave enough memory to the interface with the (\code{maxmem} parameter).
}
\examples{
\dontrun{
gui()
}
}

