% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot.model.R
\docType{methods}
\name{plot.model}
\alias{plot,SDM,ANY-method}
\alias{plot,Stacked.SDM,ANY-method}
\alias{plot.model}
\title{Plot SDMs, ensemble SDMs, and SSDMs}
\usage{
\S4method{plot}{Stacked.SDM,ANY}(x, y, ...)

\S4method{plot}{SDM,ANY}(x, y, ...)
}
\arguments{
\item{x}{Object to be plotted (S4 Algorithm.SDM, Ensemble.SDM or Stacked.SDM
object).}

\item{y,...}{Plot-based parameter not used.}
}
\value{
Open a window with a shiny app rendering all the results (habitat
 suitability map, binary map, evaluation table, variable importance and/or
 between-algorithm variance map, and/or algorithm evaluation,  and/or
 algorithm correlation matrix and/or local species richness map) in a
 user-friendly interface.
}
\description{
Allows to plot S4 \linkS4class{Algorithm.SDM}, \linkS4class{Ensemble.SDM} and
\linkS4class{Stacked.SDM} class objects.
}

