% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project.R
\docType{methods}
\name{project}
\alias{project}
\alias{project,Algorithm.SDM-method}
\alias{project,MAXENT.SDM-method}
\alias{project,Ensemble.SDM-method}
\alias{project,Stacked.SDM-method}
\title{Project model into environment}
\usage{
project(obj, Env, ...)

\S4method{project}{Algorithm.SDM}(obj, Env, ...)

\S4method{project}{MAXENT.SDM}(obj, Env, ...)

\S4method{project}{Ensemble.SDM}(obj, Env, uncertainty = TRUE,
  update.projection = TRUE, minimal.outputs = FALSE, cores = 0, ...)

\S4method{project}{Stacked.SDM}(obj, Env, method = NULL,
  uncertainty = TRUE, update.projection = TRUE,
  minimal.outputs = FALSE, cores = 0, ...)
}
\arguments{
\item{obj}{Object of class Algorithm.SDM, Ensemble.SDM or Stacked.SDM. Model(s) to be projected.}

\item{Env}{Raster stack. Updated environmental rasters to be used for projection.}

\item{...}{Additional arguments for internal use.}

\item{uncertainty}{logical. If set to TRUE, generates an uncertainty map and if update.projection is TRUE
additionally an algorithm correlation matrix.}

\item{update.projection}{logical. If TRUE, the projected rasters will replace the existing ones within the supplied .SDM object.}

\item{minimal.outputs}{logical. If TRUE, the 'sdms' slot will not contain any rasters (for memory saving purposes).}

\item{cores}{integer. Specify the number of CPU cores used to do the
computing. You can use \code{\link[parallel]{detectCores}}) to automatically
use all the available CPU cores.}

\item{method}{character. Define the method used to create the local species
richness map (for details see \code{\link[SSDM]{stack_modelling}}). If NULL (default), the method used for building the SSDM is used.}
}
\value{
Either returns the original .SDM object with updated projection slots or if minimal.outputs = TRUE only returns the projections as Raster* objects. Depending on the object class this may be: a raster (Algorithm.SDM), a raster stack (Ensemble.SDM), a biodiversity map/mean raster (Stacked.SDM).
}
\description{
This is a collection of methods to project SDMs, ESDMs or SSDMs into the supplied environment. The function is used internally to calculate the input for the projection slot of .SDM classes but can also be used to project existing .SDM objects (see Details).
}
\details{
The function uses any S4 .SDM class object and a raster stack of environmental layers of the variables the model was trained with.
}
