% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SSDM.R
\docType{package}
\name{SSDM}
\alias{SSDM}
\alias{SSDM-package}
\title{SSDM: Stacked species distribution modelling}
\description{
SSDM is a package to build multi-species distribution models. You can create
SDMs using a single or multiple algorithms (ensemble SDMs) with or without
repetition. For each species, an SDM can yield a habitat suitability map, a
binary map, an uncertainty map, and can asses variable importance, algorithm
accuracy, and between-algorithm correlation.
}
\details{
The SSDM package also provides a user-friendly interface (\code{\link{gui}}).

SSDM provides four categories of functions (that you can find in details
below): Data preparation, Modelling main functions, Model main methods, Model
classes, and Miscellaneous.
}
\section{Data preparation}{
 \describe{\item{\code{\link{load_occ}}}{Load
  occurrence data} \item{\code{\link{load_var}}}{Load environmental data}}
}

\section{Modelling main  functions}{
 \describe{
  \item{\code{\link{modelling}}}{Build an SDM using one algorithm.}
  \item{\code{\link{ensemble_modelling}}}{Build an SDM that assembles
  multiple algorithms.} \item{\code{\link{stack_modelling}}}{Build an SSDMs
  that assembles multiple algorithms and species.}}
}

\section{Model main methods}{
 \describe{
  \item{\code{\link{ensemble,Algorithm.SDM-method}}}{Build an ensemble SDM}
  \item{\code{\link{stacking,Ensemble.SDM-method}}}{Build an SSDM}
  \item{\code{\link{update,Stacked.SDM-method}}}{Update a previous SSDM with
  new occurrence data}}
}

\section{Model classes}{
 \describe{
  \item{\code{\linkS4class{Algorithm.SDM}}}{S4 class to represent SDMs}
  \item{\code{\linkS4class{Ensemble.SDM}}}{S4 class to represent ensemble
  SDMs} \item{\code{\linkS4class{Stacked.SDM}}}{S4 class to represent SSDMs}}
}

\section{Miscellaneous}{
 \describe{ \item{\code{\link{gui}}}{User-friendly
  interface for SSDM package} \item{\code{\link{plot.model}}}{Plot SDMs}
  \item{\code{\link{save.model}}}{Save SDMs}
  \item{\code{\link{load.model}}}{Load SDMs}}
}

