% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacking.R
\docType{methods}
\name{stacking}
\alias{stacking}
\alias{stacking,Ensemble.SDM-method}
\title{Stack different ensemble SDMs in one SSDM}
\usage{
stacking(enm, ..., name = NULL, method = "P", metric = "SES",
  thresh = 1001, rep.B = 1000, verbose = T, GUI = F)

\S4method{stacking}{Ensemble.SDM}(enm, ..., name = NULL, method = "P",
  metric = "SES", thresh = 1001, rep.B = 1000, verbose = T, GUI = F)
}
\arguments{
\item{enm,...}{character. Ensemble SDMs to be stacked.}

\item{name}{character. Optinnal name given to the final SSDM producted (by
default 'Species.SDM').}

\item{method}{character. Define the method used to create the local species
richness map (see details below).}

\item{metric}{character. Metric used to compute the binary map threshold (see
details below.)}

\item{thresh}{numeric. A single integer value representing the number of equal
interval threshold values between 0 & 1. The higher it is the more accurate
is the threshold but the longer is the modelling evaluation step (see
\code{\link[SDMTools]{optim.thresh}}).}

\item{rep.B}{integer. If the method used to create the local species richness
is random bernoulli (\strong{B}), rep.B parameter defines the number of
repetition used to create random bernoulli binary maps for each species.}

\item{verbose}{logical. If true allow the function to print text in the
console.}

\item{GUI}{logical. Don't take that argument into account (parameter for the
 user interface).}
}
\value{
an S4 \linkS4class{Stacked.SDM} class object viewable with
 \code{\link{plot.model}} function.
}
\description{
This is a function to stack several ensemble SDMs in one SSDM. The function
takes in inputs several S4 \linkS4class{Ensemble.SDM} class objects produced
with \code{\link{ensemble_modelling}} or \code{\link{ensemble}} functions. The
function returns an S4 \linkS4class{Stacked.SDM} class object containing the
local species richness map, and the uncertainty map based on the habitat
suitability map variance inter algorithm, all evaluation tables comming with
(model evaluation, algorithm evaluation, algorithm correlation matrix and
variable importance), and all associated ensemble SDMs for each species (see
\code{\link{ensemble_modelling}}).
}
\details{
\strong{Metric :} choice of the metric used to compute binary map
 threshold and confusion matrix (by default SES as recommanded by Liu et al.
 2005,see references below): \describe{ \item{"Kappa"}{maximizes the model
 Kappa value} \item{"TSS"}{\strong{True Skill Statistic} maximizes the
 sensitivity and specificity sum} \item{"CCR"}{maximizes the correct
 predicted observations proportion} \item{"SES"}{using the sensitivty
 specificity equality} \item{"LW"}{using the lowest occurence prediction
 probability} \item{"ROC"}{minimizing the distance between the ROC plot
 (receiving poerating characteristic curve) and the upper left coin (1,1)} }

 \strong{Methos :} Choice of the method used to compute the local species
 richness map (see Calabrez et al. 2014 for more informations, see references
 below) : \describe{\item{P}{(Probablity) sum probabilities of the habitat
 suitability maps }\item{B}{(Random bernoulli) drawing repeatedly from a
 Bernoulli distribution}\item{T}{(Threshold) sum the binary map obtained with
 the thresholding (depending of the metric, see metric parameter).}}
}
\examples{
\dontrun{
stacking(Specie1.enm, Specie2.enm)
}
}
\references{
Liu,  C.  et  al.  2005.  Selecting  thresholds  of  occurrence in
 the prediction  of  species  distributions./ Ecography  28:  385 / 393.

 Calabrese, J.M., Certain, G., Kraan, C. & Dormann, C.F. (2014) Stacking
 species  distribution  models  and  adjusting  bias  by linking them to
 macroecological models. Global Ecology and Biogeography, 23, 99-112.
}
\seealso{
\code{\link{stack_modelling}} for stack species ensemble modelling
 with multiple algorithms and multiples species
}

