% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/stacking.R
\docType{methods}
\name{stacking}
\alias{stacking}
\alias{stacking,Ensemble.SDM-method}
\title{Stack different ensemble SDMs in an SSDM}
\usage{
stacking(enm, ..., name = NULL, method = "P", rep.B = 1000,
  range = NULL, endemism = c("WEI", "Binary"), verbose = T, GUI = F)

\S4method{stacking}{Ensemble.SDM}(enm, ..., name = NULL, method = "P",
  rep.B = 1000, range = NULL, endemism = c("WEI", "Binary"),
  verbose = T, GUI = F)
}
\arguments{
\item{enm,...}{character. Ensemble SDMs to be stacked.}

\item{name}{character. Optional name given to the final SSDM produced (by
default 'Species.SDM').}

\item{method}{character. Define the method used to create the local species
richness map (see details below).}

\item{rep.B}{integer. If the method used to create the local species richness
is the random bernoulli (\strong{B}), rep.B parameter defines the number of
repetitions used to create binary maps for each species.}

\item{range}{integer. Set a value of range restriction (in pixels) around
presences occurrences on habitat suitability maps (all further points will
have a null probability, see Crisp et al (2011) in references). If NULL, no
range restriction will be applied.}

\item{endemism}{character. Define the method used to create an endemism map
(see details below).}

\item{verbose}{logical. If set to true, allows the function to print text in
the console.}

\item{GUI}{logical. Don't take that argument into account (parameter for the
 user interface).}
}
\value{
an S4 \linkS4class{Stacked.SDM} class object viewable with the
 \code{\link{plot.model}} function.

 \strong{Methods:} Choice of the method used to compute the local species
 richness map (see Calabrez et al. (2014) for more informations, see
 reference below): \describe{\item{P}{(Probablity) sum probabilities of
 habitat suitability maps }\item{B}{(Random bernoulli) draw repeatedly from a
 Bernoulli distribution}\item{T}{(Threshold) sum the binary map obtained with
 the thresholding (depending on the metric, see metric parameter).}}

 \strong{Endemism:} Choice of the method used to compute the endemism map
 (see Crisp et al. (2001) for more information, see reference below):
 \describe{\item{NULL}{No endemism map}\item{WEI}{(Weighted Endemism Index)
 Endemism map built by counting all species in each cell and weighting each
 by the inverse of its range} \item{CWEI}{(Corrected Weighted Endemism Index)
 Endemism map built by dividing the weighted endemism index by the total
 count of species in the cell.}}First string of the character is the method
 either WEI or CWEI, and in those cases second string of the vector is used
 to precise range calculation, whether the total number of occurrences
 \strong{'NbOcc'} whether the surface of the binary map species distribution
 \strong{'Binary'}.
}
\description{
This is a function to stack several ensemble SDMs in an SSDM. The function
takes as inputs several S4 \linkS4class{Ensemble.SDM} class objects produced
with \code{\link{ensemble_modelling}} or \code{\link{ensemble}} functions. The
function returns an S4 \linkS4class{Stacked.SDM} class object containing the
local species richness map, the between-algorithm variance map, and all
evaluation tables coming with (model evaluation, algorithm evaluation,
algorithm correlation matrix and variable importance), and a list of ensemble
SDMs for each species (see \code{\link{ensemble_modelling}}).
}
\examples{
\dontrun{
# Loading data
data(Env)
data(Occurrences)
Occ1 = subset(Occurrences, Occurrences$SPECIES == 'elliptica')
Occ2 = subset(Occurrences, Occurrences$SPECIES == 'gracilis')

# SSDM building
ESDM1 = ensemble_modelling(c('CTA', 'SVM'), Occ1, Env, rep = 1,
                           Xcol = 'LONGITUDE', Ycol = 'LATITUDE',
                           name = 'elliptica', ensemble.thresh = c(0.6))
ESDM2 = ensemble_modelling(c('CTA', 'SVM'), Occ2, Env, rep = 1,
                           Xcol = 'LONGITUDE', Ycol = 'LATITUDE',
                           name = 'gracilis', ensemble.thresh = c(0.6))
SSDM = stacking(ESDM1, ESDM2)

# Results plotting
plot(SSDM)
}
}
\references{
C. Liu, P. M. Berry, T. P. Dawson,  R. & G. Pearson (2005)
 "Selecting thresholds of occurrence in the prediction of species
 distributions." \emph{Ecography} 28:85-393
 \url{http://www.researchgate.net/publication/230246974_Selecting_Thresholds_of_Occurrence_in_the_Prediction_of_Species_Distributions}


























 J.M. Calabrese, G. Certain, C.  Kraan, & C.F. Dormann (2014) "Stacking
 species distribution  models  and  adjusting  bias  by linking them to
 macroecological models." \emph{Global Ecology and Biogeography} 23:99-112
 \url{http://portal.uni-freiburg.de/biometrie/mitarbeiter/dormann/calabrese2013globalecolbiogeogr.pdf}









 M. D. Crisp, S. Laffan, H. P. Linder & A. Monro (2001) "Endemism in the
 Australian flora"  \emph{Journal of Biogeography} 28:183-198
 \url{http://biology-assets.anu.edu.au/hosted_sites/Crisp/pdfs/Crisp2001_endemism.pdf}
}
\seealso{
\code{\link{stack_modelling}} to build SSDMs.
}

